Update Babel to version 6.
Babel 5 had nonobvious behavior where if you had exactly one default export,
it would automatically add a `module.exports = exports["default"];` statement
at the end. This is normally fine, but would break certain transpilers (like
Webpack) by overwriting the `exports` object when it expects an ES6-style export.

https://github.com/babel/babel.github.io/blob/862b43db93e48762671267034a50c30c00e433e2/docs/usage/modules.md#interop

By removing the export of `timers`, I changed the export semantics accidentally.
I couldn't find a way to disable this feature in Babel 5, so I just upgraded to
Babel 6 instead, which does not include it (though there is a plugin for it).